Replace ToolPolicy enum with ToolConstraints product type
ToolPolicy<T> mixed availability and requirement into a single sum type
with shapes that couldn't be lowered losslessly to all adapters. Replace
it with a product type exposing only the adapter-exact portable subset:

- ToolAvailability<S>: All | Only(Vec<S>)
- ToolRequirement<S>: Optional | AtLeastOne | Specific(S)
- ToolConstraints<T>: { available, requirement }

Builder API changes:
- tools::<T>() now defaults to All + Optional (was Disabled)
- available_tools([..]) replaces allow_only / allow_all
- require_any_tool() replaces require_all
- require_tool(s) replaces require_only([s])

Lowering rules in erase_turn_config:
- Optional → Auto, AtLeastOne → Required, Specific(s) → Specific
- Empty tool list always produces AdapterToolChoice::None regardless
  of requirement, so available_tools([]) is correctly disabled at
  the adapter boundary
- require_tool(x) where x ∉ Only([..]) returns AgentError::InvalidToolConstraints
- require_any_tool() with empty resolved list returns the same error

+++ demo-apps/sqlite/agent/src/db/executor.rs
@@ -259,7 +259,7 @@ impl SqliteDb {
     pub fn execute_write(&self, sql: &str) -> Result<ModifyResult, DbError> {
         let conn = self.conn.lock().unwrap();
         conn.execute_batch(sql)?;
-        let rows_affected = conn.changes() as u64;
+        let rows_affected = conn.changes();
         Ok(ModifyResult {
             rows_affected,
             message: format!("{rows_affected} row(s) affected"),
